web_scraping/pyproject.toml
[project]
name = "fastboard-web-scraping"
version = "0.1.0"
description = "Web scraping service using Crawl4AI for FastboardAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "crawl4ai>=0.4.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
web-scraping = "web_scraping.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/web_scraping"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
